Fixed issue where recipe description reverted to basic verb description on start
StartingSlotsUpdated was called by elements being moved from the starting slots to storage. This meant the window behaved as if it was unstarted and starting slots were being fiddled with. I've called Situation.Start() before moving anything to storage, and added a check to avoid doing any of the Notes text updates unless the situation is unstarted