Ensure working with Health will spawn the Time verb
Since Health is not marked a "job", it would not spawn Time upon
completion, and we can't rely on inductions in this case anyway because
Health is not always just worked.

The solution is therefore to make the initial "safe" versions of menial
labor always spawn the Time verb on first run through an additional
recipe.