Add some additional mechanisms for handling corrupt saves
Save loads are now more often wrapped in exception catchers to handle
unknown problems, which should lower the chance of getting a black screen of
death.

Additionally, saving is now first performed to a temporary file, which is then
validated, and finally renamed if the save appears to have been successful. This
happens ON TOP of the existing backup mechanisms.